refactor: demonstrate desired near-api behavior
THIS DOESN'T WORK; IT IS FOR DEMONSTRATION PURPOSES ONLY

This change demonstrates a proposed modification to `near-api-js`.
Notable changes:

* Rename library to `near-api`, since it being served through `npm`
  already indicates that it's a JS lib.

* Remove `src/config.js` from project; these values are defaults that
  can be retrieved from `near-api` directly based on the current
  NODE_ENV (`testnet` for `NODE_ENV=development`; and eventually
  `mainnet` for `NODE_ENV=production`). If a project needs to
  customize certain parameters, `near-api` can provide a way to do so.
  This project would not need to customize any parameters, so such a
  hypothetical interface is not documented here.

* Now that `src/config.js` is gone, we can also have `near-api` use a
  sensible `keyStore` default value based on where it's being run. This
  app runs in the browser; `BrowserLocalStorageKeyStore` should be the
  default. Now initializing a `near` object at the beginning of
  `index.js` is no longer necessary.

* Make the `near-api` interface closely match the `near-shell` interface
  proposed in https://github.com/nearprotocol/NEPs/pull/31. Here are the
  interfaces demonstrated here:

  * `near.account.login`
  * `near.account.logout`
  * `near.account.state`
  * `near.contract.call`
  * `near.contract.view`

As you can see, the proposed changes cut out a tremendous amount of
boilerplate. And following the suggestions in NEP-31 in `near-api` as
well as `near-shell` will help developers in both of these contexts
develop better mental models for what they can do with NEAR ("oh, this
is something related to _accounts_"; "this is something related to a
_contract_").

For parity with the old Contract interface, this change adds a new
`Contract` class. This class is a simple wrapper around the raw calls to
`near.contract.view` and `near.contract.call`. Eventually, such a class
could be automatically generated based on a contract's Wasm binary
(@willemneal has ideas for this) using a command like `near codegen`.
But for now, having project authors maintain their own similar classes,
if they care to, doesn't seem like much more effort than the old
approach.

A huge benefit of the new approach to `contract.call` that isn't super
obvious: *no magic*. The old `new Contract` approach created some
[dynamically-generated functions][near-api-js] like
`contract.addMessage`. But these dynamically-generated functions come
with a huge downside: no inline documentation! That is, when using a
modern development environment such as VS Code or a project & editor
configured with TypeScript, the documentation from `near-api-js` shows
up as you type. This is an incredible productivity tool! We want to push
this as the default experience for NEAR newbies. But
dynamically-generated functions can't have statically-generated
documentation. This new approach avoids that, so when someone types
`near.contract.call`, they'll see that they can attach an optional `gas`
and `amount` value, without needing to leave their editor to find this
information.

  [near-api-js]: https://github.com/near/near-api-js/blob/790deaba22c59261a6d561694368036c95b928ac/src/contract.ts#L38